+++ config/runtime.exs
@@ -64,3 +64,6 @@ config :glific, Glific.Communications.Mailer,
   region: env!("SES_REGION", :string!, "This is not a Region"),
   access_key: env!("SES_KEY", :string!, "This is not a Key"),
   secret: env!("SES_SECRET", :string!, "This is not a Secret")
+
+config :glific,
+  google_captcha_secret_key: env!("RECAPTCHA_SECRET_KEY", :string!, "This is not a Secret")
